# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post API",
        "description": "Simple API documentation for providing full posts data",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns paginated list of posts",
                "operationId": "getPostsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{uuid}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by UUID",
                "description": "Returns a single post",
                "operationId": "getPostByUuid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "8 symbols UUID of the post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "c2d33f86"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "title": "Post",
                "required": [
                    "uuid",
                    "title",
                    "content"
                ],
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "abc123ef"
                    },
                    "title": {
                        "type": "string",
                        "example": "Catchy title"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is the full content of the post with Markdown."
                    },
                    "featured_image_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "dislikes_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 4
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T15:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Posts",
            "description": "Operations about blog posts"
        }
    ]
}